requirements.txt
fastapi
uvicorn
httpx
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
